Still figuring out how would Cypress API allow various calls
